- modified gitignore - further dev on svg elements and templates

+++ Uebung_3/src/gp2.svg-generator/src/main/java/gp2/templating/shape/api/Coordinate.java
@@ -14,13 +14,13 @@ import java.util.Objects;
 public class Coordinate {
 
     @Getter
-    private final int x;
+    private final double x;
 
     @Getter
-    private final int y;
+    private final double y;
 
-    public Coordinate(int x,
-                      int y) {
+    public Coordinate(double x,
+                      double y) {
         this.x = x;
         this.y = y;
     }
